Change args to `get_analytics`
We now only pass the S.No of the entry whose analytics we wish to
find.
It is expected that when the user makes a search, they choose one
entry from a list of matching entries, and that entry has the S.No
attached to it, which will be used to get the analytics here.