Add endpoint for searching
We use pandas to search relevant rows in the data,
and expose an endpoint to search compnaies based on company name and country code. Partial matching entries are shown too,
both for the company name and the country code. If multiple entries
match the search filters, they are all listed, and can be differentiated
by their serial number (which is unique for each entry in the
data).